修改计算当前毫秒的计算方式

+++ jni/date/time.h
@@ -24,7 +24,7 @@ class Time {
      * 可用来计算某代码块运行时间时使用
      * @return
      */
-    static int64_t currentTimeMillis();
+    static double currentTimeMillis();
 };
 
 NS_GODIN_END
